Video-LLaMA/requirement.txt
tqdm
decord
timm
einops
opencv_python
torchvision
pytorch 
torchaudio 
pytorch-cuda
bitsandbytes
omegaconf
iopath
webdataset
ftfy
SentencePiece
transformers==4.28.0
gradio==3.24.1
pytorchvideo
gradio-client==0.0.8
